[SW-107][minor_changed]홈페이지 컴포넌트 분리
- 홈 페이지의 컴포넌트 소스코드 분리 및 이동
