+++ front/src/components/Descript.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+function Descript(): JSX.Element {
+  return (
+    <>
+      무료로 친구들과 즐거운 대화를 나눠보세요.
+      <br />
+      VOICE SPACE는 피로하지 않은 원격 소통 서비스를 제공합니다.
+      <br />
+    </>
+  );
+}
+
+export default Descript;
