+++ front/src/components/MoreInfo.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+function MoreInfo(): JSX.Element {
+  return (
+    <>
+      <br />
+      VOICE SPACE에 대해
+      <a href="www.naver.com"> 자세히 알아보세요</a>
+    </>
+  );
+}
+
+export default MoreInfo;
