+++ front/src/pages/homePage/home.tsx
@@ -1,112 +1,13 @@
-import React, {useState} from 'react';
+import React from 'react';
 import {RouteComponentProps} from 'react-router-dom';
-import {message, Button} from 'antd';
+import {message} from 'antd';
 import {v4 as uuidV4} from 'uuid';
 import './home.css';
-
-function Welcome(): JSX.Element {
-  return (
-    <>
-      <img className="logoImage" src="./assets/home/homeLogo.png" />
-      <div className="welcome">
-        WELCOME
-        <br />
-        VOICE SPACE
-      </div>
-    </>
-  );
-}
-
-function Descript(): JSX.Element {
-  return (
-    <>
-      무료로 친구들과 즐거운 대화를 나눠보세요.
-      <br />
-      VOICE SPACE는 피로하지 않은 원격 소통 서비스를 제공합니다.
-      <br />
-    </>
-  );
-}
-
-interface CreateRoomProps {
-  createRoomButtonClick: () => void;
-}
-
-function CreateRoom(props: CreateRoomProps) {
-  return (
-    <div>
-      <br />
-      <Button
-        id="button"
-        shape="round"
-        type="primary"
-        onClick={props.createRoomButtonClick}
-      >
-        새로운 보이스 채팅
-      </Button>
-    </div>
-  );
-}
-
-interface EnterRoomProps {
-  enterRoomButtonClick: (arg0: string) => void;
-}
-
-function EnterRoom(props: EnterRoomProps): JSX.Element {
-  const [roomId, setRoomId] = useState('');
-
-  // const info = () => {
-  //   message.info('유효하지 않은 코드입니다');
-  // };
-  const element: JSX.Element = (
-    <>
-      <br />
-      <Button id="button" onClick={() => props.enterRoomButtonClick(roomId)}>
-        참가하기
-      </Button>
-    </>
-  );
-
-  const [focus, setFocus] = useState(0);
-
-  const onFocus = () => {
-    setFocus(1);
-  };
-
-  const onBlur = (e: React.FocusEvent<HTMLInputElement>) => {
-    if (e.target.value === '') setFocus(0);
-  };
-
-  const inputOnchange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setRoomId(e.target.value);
-  };
-
-  return (
-    <div>
-      또는
-      <br />
-      <input
-        className="input"
-        onFocus={onFocus}
-        onBlur={onBlur}
-        placeholder="코드(RoomId)를 입력해주세요."
-        value={roomId}
-        onChange={inputOnchange}
-      />
-      {focus ? element : null}
-    </div>
-  );
-}
-
-function MoreInfo(): JSX.Element {
-  return (
-    <>
-      <br />
-      VOICE SPACE에 대해
-      <a href="www.naver.com"> 자세히 알아보세요</a>
-    </>
-  );
-}
+import Welcome from '../../components/Welcome';
+import Descript from '../../components/Descript';
+import CreateRoom from '../../components/CreateRoom';
+import EnterRoom from '../../components/EnterRoom';
+import MoreInfo from '../../components/MoreInfo';
 
 function Home(props: RouteComponentProps): JSX.Element {
   const enterRoomClick = (roomId: string) => {
@@ -123,7 +24,12 @@ function Home(props: RouteComponentProps): JSX.Element {
   return (
     <div className="home">
       <div className="top">
-        <Welcome />
+        <Welcome
+          logoImageSrc="./assets/home/homeLogo.png"
+          logoImageOnload={() => {
+            console.log('logoImageOnload called');
+          }}
+        />
       </div>
       <div className="bottom">
         <Descript />
